requirements.txt
uvicorn~=0.30.1
fastapi~=0.103.0
pydantic~=2.7.3
python-dotenv~=1.0.1
loguru~=0.7.2
halo~=0.0.31
requests~=2.32.2
openai~=1.30.5